+++ packs/alertlogic/actions/lib/get_scan_results.py
@@ -14,6 +14,7 @@
 # limitations under the License.
 
 import requests
+from lib.base import API_HOST
 
 
 def booleen2string(booleen):
@@ -33,7 +34,8 @@ def GetScanResults(config, scan_exec_id, new_vulns=False, new_ports=False):
     ValueError: On lack of key in config.
     """
 
-    url = "https://{}/api/scan/v1/results/{}".format(config['api_host'], scan_exec_id)
+    url = "https://{}/api/scan/v1/results/{}".format(API_HOST,
+                                                     scan_exec_id)
 
     payload = {}