+++ packs/duo/tests/test_action_auth_ping.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python
-
 # Licensed to the StackStorm, Inc ('StackStorm') under one or more
 # contributor license agreements.  See the NOTICE file distributed with
 # this work for additional information regarding copyright ownership.
@@ -13,26 +11,12 @@
 # distributed under the License is distributed on an "AS IS" BASIS,
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
-# limitations under the License.
 
-from lib.actions import AuthBaseAction
+from duo_base_test_case import DuoBaseActionTestCase
+
+from auth_ping import AuthPingAction
 
 
-class AuthCheckAction(AuthBaseAction):
-    def run(self):
-        """
-        Ping the Duo Platorm.
-
-        Returns: An dict with info returned by Duo.
-
-        Raises:
-          RuntimeError: On Invalid keys.
-        """
-
-        try:
-            data = self.duo_auth.check()
-        except RuntimeError, e:
-            self.send_user_error(e)
-            raise RuntimeError("Check failed! '%s'" % e)
-        else:
-            return data
+class AuthPingActionTestCase(DuoBaseActionTestCase):
+    __test__ = True
+    action_cls = AuthPingAction