+++ packs/alertlogic/CHANGES.md
@@ -4,6 +4,10 @@
 
 * n/a
 
+## 0.2
+
+- Rename `config.yaml` to `config.schema.yaml` and update to use schema.
+
 ## 0.1
 
 * Initial release.