+++ packs/opsgenie/CHANGES.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.2
+
+- Rename `config.yaml` to `config.schema.yaml` and update to use schema.
+
 ## 0.1
 
 - First release.